+++ src/ast/expr/int_lit.h
@@ -3,12 +3,14 @@
 
 #include "ast/expr/expr.h"
 
+#include <stdint.h>
+
 typedef struct ast_int_lit
 {
     ast_expr_t base;
-    int value;
+    int64_t value;
 } ast_int_lit_t;
 
-ast_expr_t* ast_int_lit_create(int value);
+ast_expr_t* ast_int_lit_create(int64_t value);
 
 #endif
